lucetc: move compiler to own module
